Float instead of double

+++ src/com/mnenmenth/firenfrost/entity/AnimFrame.cpp
@@ -7,13 +7,13 @@
 #include "AnimFrame.h"
 #include <stdexcept>
 
-AnimFrame::AnimFrame(Viewport::Dimension<double> size, sf::IntRect sheet_rect) {
+AnimFrame::AnimFrame(Viewport::Dimension<float> size, sf::IntRect sheet_rect) {
     setSize(size);
     defaultHitbox = Viewport::pixelToCoord(sheet_rect);
 }
 
-AnimFrame::AnimFrame(double width, double height, int sheet_frameWidth, int sheet_frameHeight, int sheet_frameTop, int sheet_frameLeft) {
-    AnimFrame(Viewport::Dimension<double>{width, height}, sf::IntRect(sheet_frameWidth, sheet_frameHeight, sheet_frameTop, sheet_frameLeft));
+AnimFrame::AnimFrame(float width, float height, int sheet_frameWidth, int sheet_frameHeight, int sheet_frameTop, int sheet_frameLeft) {
+    AnimFrame(Viewport::Dimension<float>{width, height}, sf::IntRect(sheet_frameWidth, sheet_frameHeight, sheet_frameTop, sheet_frameLeft));
 }
 
 void AnimFrame::setFrameTime(float time) {
@@ -24,58 +24,58 @@ float AnimFrame::getFrameTime() {
     return frameTime;
 }
 
-void AnimFrame::setSize(double width, double height) {
-    setSize(Viewport::Dimension<double>{width, height});
+void AnimFrame::setSize(float width, float height) {
+    setSize(Viewport::Dimension<float>{width, height});
 }
 
-void AnimFrame::setSize(Viewport::Dimension<double> size) {
+void AnimFrame::setSize(Viewport::Dimension<float> size) {
     this->size = size;
     setScale((float)(this->size.toPixel().width/sheet_frameBounds.width), (float)(this->size.toPixel().height/sheet_frameBounds.height));
 }
 
-double AnimFrame::getWidth() {
+float AnimFrame::getWidth() {
     return size.width;
 }
 
-void AnimFrame::setWidth(double width) {
+void AnimFrame::setWidth(float width) {
     size.width = width;
     setScale((float)(size.toPixel().width/sheet_frameBounds.width), 1.0f);
 }
 
-double AnimFrame::getHeight() {
+float AnimFrame::getHeight() {
     return size.height;
 }
 
-void AnimFrame::setHeight(double height) {
+void AnimFrame::setHeight(float height) {
     size.height = height;
     setScale(1.0f, (float)(size.toPixel().height/sheet_frameBounds.height));
 }
 
-Viewport::Dimension<double> AnimFrame::getSize() {
+Viewport::Dimension<float> AnimFrame::getSize() {
     return size;
 }
 
-void AnimFrame::setRelativePos(Viewport::Point<double> pos) {
+void AnimFrame::setRelativePos(Viewport::Point<float> pos) {
     relativePos = pos;
 }
 
-void AnimFrame::setRelativeX(double x) {
+void AnimFrame::setRelativeX(float x) {
     relativePos.x = x;
 }
 
-void AnimFrame::setRelativeY(double y) {
+void AnimFrame::setRelativeY(float y) {
     relativePos.y = y;
 }
 
-Viewport::Point<double> AnimFrame::getRelativePos() {
+Viewport::Point<float> AnimFrame::getRelativePos() {
     return relativePos;
 }
 
-void AnimFrame::setRelativeRot(double rot) {
+void AnimFrame::setRelativeRot(float rot) {
     relativeRot = rot;
 }
 
-double AnimFrame::getRelativeRot() {
+float AnimFrame::getRelativeRot() {
     return relativeRot;
 }
 
@@ -95,7 +95,7 @@ sf::IntRect AnimFrame::getSheet_frameBounds() {
     return sheet_frameBounds;
 }
 
-void AnimFrame::addHitbox(std::string name, sf::Rect<double> hitbox) {
+void AnimFrame::addHitbox(std::string name, sf::Rect<float> hitbox) {
     hitboxes[name] = hitbox;
 }
 
@@ -107,7 +107,7 @@ void AnimFrame::removeHitbox(std::string name) {
         throw std::invalid_argument(name + " is not a hitbox");
 }
 
-sf::Rect<double>& AnimFrame::getHitbox(std::string name) {
+sf::Rect<float>& AnimFrame::getHitbox(std::string name) {
     auto entry = hitboxes.find(name);
     if (entry != hitboxes.end())
         return hitboxes[name];
